requirements.txt
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
httpx==0.25.2
python-telegram-bot==20.7
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1